refactor : baseEntity 위치 변경
