+++ src/main/java/dbdr/domain/institution/entity/Institution.java
@@ -1,6 +1,6 @@
 package dbdr.domain.institution.entity;
 
-import dbdr.domain.core.entity.BaseEntity;
+import dbdr.domain.core.base.entity.BaseEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Table;
